requirements.txt
bing-image-downloader
pillow
scikit-image
numpy
scikit-learn
